+++ apis/profile/apis.ts
@@ -1,4 +1,5 @@
-import { ProfileResponse } from './type';
 import privateApi from '../config/privateApi';
+
+import type { ProfileResponse } from './type';
 
 export const getProfile = () => privateApi.get<ProfileResponse>('/me/page');
