+++ components/common/Spacing/index.ts
@@ -1,2 +1,3 @@
 export { default as Spacing } from './Spacing';
 export { default as DivisionSpacing } from './DivisionSpacing';
+export { default as DivisionBar } from './DivisionBar';
